requirements_no_mono.txt
SoundFile
torchaudio==2.7.0
munch
torch==2.7.0
pydub
pyyaml
librosa
nltk
matplotlib
accelerate==0.31.0
transformers==4.52.3
einops
einops-exts
tqdm
typing
typing-extensions
